# backend/docs/swagger.yaml
basePath: /
definitions:
  docs.ErrorResponse:
    properties:
      error:
        example: Message d'erreur
        type: string
    type: object
  dtos.CategoryRequest:
    description: Informations catégorie pour création/modification
    properties:
      name:
        description: Nom de la catégorie
        example: Visage
        type: string
    required:
    - name
    type: object
  dtos.CategoryResponse:
    description: Informations catégorie
    properties:
      createdAt:
        description: Date de création
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        description: UUID de la catégorie
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        description: Nom de la catégorie
        example: Visage
        type: string
      updatedAt:
        description: Date de mise à jour
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  dtos.PatchCategoryRequest:
    description: Permet de mettre à jour uniquement le nom d'une catégorie
    properties:
      name:
        description: Nom de la catégorie (optionnel)
        example: Visage
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: "API REST complète pour une boutique de produits de soins (skincare
    shop) avec authentification JWT et gestion des rôles.\nCette API permet de gérer
    les produits, catégories, commandes et utilisateurs.\n\n## Authentification\nLa
    plupart des endpoints nécessitent une authentification JWT. Pour tester les endpoints
    protégés :\n\n**Étape 1 : Obtenir un token**\n1. Utilisez `/auth/register` pour
    créer un compte OU `/auth/login` pour vous connecter\n2. Copiez le token JWT retourné
    dans la réponse (champ \"token\")\n\n**Étape 2 : Autoriser les requêtes dans Swagger
    UI**\n1. Cliquez sur le bouton **\"Authorize\"** (icône de cadenas) en haut à
    droite de cette page\n2. Dans le champ \"Value\", entrez : `Bearer <votre-token>`
    (remplacez `<votre-token>` par le token copié)\n3. Cliquez sur **\"Authorize\"**
    puis **\"Close\"**\n\n**Étape 3 : Tester les endpoints**\nTous les endpoints protégés
    sont maintenant accessibles. Vous pouvez cliquer sur \"Try it out\" et tester
    chaque endpoint.\n\n## Rôles\n- **USER** : Peut voir les produits, créer des commandes,
    voir ses propres commandes\n- **ADMIN** : Accès complet à toutes les ressources\n\n---\n\n##
    \U0001F510 Compte de test Admin\n\n**Email:** `momo@ynov.com`\n**Password:** `Password2025`\n\nUtilisez
    ce compte pour tester toutes les fonctionnalités admin dans l'API."
  title: PORELO Skincare Shop API
...
paths:
  /admin/categories:
    get:
      consumes:
      - application/json
      description: Récupère la liste de toutes les catégories (admin uniquement)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CategoryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Liste toutes les catégories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Crée une nouvelle catégorie (admin uniquement)
      parameters:
      - description: Informations de la catégorie
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Catégorie avec ce nom existe déjà
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Créer une catégorie
      tags:
      - Categories
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Supprime une catégorie (admin uniquement)
      parameters:
      - description: ID de la catégorie
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Supprimer une catégorie
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Récupère les détails d'une catégorie par son ID (admin uniquement)
      parameters:
      - description: ID de la catégorie
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CategoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Détails d'une catégorie
      tags:
      - Categories
    patch:
      consumes:
      - application/json
      description: 'Met à jour uniquement le nom d''une catégorie (admin uniquement).
        Exemple: changer uniquement le nom sans recréer la catégorie.'
      parameters:
      - description: ID de la catégorie
        in: path
        name: id
        required: true
        type: string
      - description: Champs à mettre à jour (nom optionnel)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.PatchCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mettre à jour partiellement une catégorie
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Met à jour une catégorie existante (admin uniquement)
      parameters:
      - description: ID de la catégorie
        in: path
        name: id
        required: true
        type: string
      - description: Nouvelles informations de la catégorie
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mettre à jour une catégorie
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" suivi d''un espace et du token JWT. Exemple: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"